test(snapshots): merge SnapshotChunkReaderTest to FileBasedSnapshotChunkReaderTest
The test was assuming lots of internal details. So it makes sense to move it to the implementation test.
